LowPassFilter changes, add External_Audio_Output

+++ examples/examples/09.Delays/AudioDelayFeedback/AudioDelayFeedback.ino
@@ -60,7 +60,7 @@ int updateAudio(){
   char asig = aTriangle.next(); // get this so it can be used twice without calling next() again
   //return asig/8 + aDel.next(asig, (uint16_t) del_samps); // mix some straight signal with the delayed signal
   //return aDel.next(aTriangle.next(), (uint16_t) del_samps); // instead of the previous 2 lines for only the delayed signal
-  return asig/8 + aDel.next(asig, deltime); // mix some straight signal with the delayed signal
+  return asig/8 + aDel.next(asig, (uint16_t) deltime); // mix some straight signal with the delayed signal
 }
